+++ packages/linux_secret_store/xdg_secret_portal_store_default/lib/xdg_secret_portal_store_default.dart
@@ -1,0 +1,5 @@
+/// A default crypto implementation of [`package:xdg_secret_portal_store`](https://pub.dev/packages/xdg_secret_portal_store)
+/// that uses [`package:cryptography`](https://pub.dev/packages/cryptography).
+library;
+
+export 'src/secret_store_crypto_default.dart';